Add credentials, extend authorization rules and set rewrites inside next.config
New configurations in the authorization module were added to send credentials with each request. The rules to check if a user is authorized were extended to include admin and static paths. Rewrites in the Next.js configuration were setup for admin and static paths to support reverse proxy setup. A link to the admin page was also added on the dashboar.