Adds a table of user files to the UI on the dashboard, pagination and data loading through the API
This commit adds a table of uploaded user files to the dashboard page. The table has the ability to paginate page by page, and also displays basic information about the file, in addition there are buttons to delete and edit the file.

In addition to the table and work with it, the table data loading and pagination processing, as well as the functions of adding and deleting files are implemented.