Move adminTable and remove admin prop from DeleteItem
The adminTable.tsx file has been moved from the 'dashboard' folder to the 'adminPanel' folder to better reflect the associated actions and contents. The 'admin' prop from the DeleteItem component in buttons.tsx was removed as it is no longer required. Also, error messages in data.ts have been slightly adjusted for clarity.