Refactor UI folder structure and delete unnecessary code files
This commit consists of two primary changes, first being renaming various components to camel case which is more in line with general coding practices. The second modification is the deletion of several files including 'cards.tsx', 'home.module.css', 'modal-upload.tsx' and 'status.tsx', which seemed to be unused and unnecessary, thereby cleaning up the codebase.